requirements.txt
# Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI / LLM (OpenAI 兼容)
openai>=1.12.0
httpx>=0.26.0
dashscope>=1.20.0

# 数据库
aiosqlite>=0.19.0

# 环境变量
python-dotenv>=1.0.0

# 知识库文件解析（PDF）
pypdf>=4.0.0

# 可选：RAG 向量检索（若需本地向量库可取消注释）
# chromadb>=0.4.22
# sentence-transformers>=2.2.0
